Refactor mock for one class
